backend/requirements.txt
annotated-types
APScheduler
asgiref
blis
cachetools
catalogue
certifi
charset-normalizer
click
cloudpathlib
confection
cymem
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
dotenv
duckdb
easyocr
filelock
fsspec
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
idna
ImageIO
Jinja2
joblib
langcodes
language_data
lazy_loader
marisa-trie
markdown-it-py
MarkupSafe
mdurl
mpmath
murmurhash
networkx
ninja
numpy
opencv-python-headless
packaging
pandas
pillow
preshed
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pyclipper
pydantic
pydantic_core
Pygments
PyJWT
pytesseract
python-bidi
python-dateutil
python-dotenv
pytz
PyYAML
requests
rich
rsa
scikit-image
scikit-learn
scipy
shapely
shellingham
six
smart_open
spacy
spacy-legacy
spacy-loggers
sqlparse
srsly
sympy
thinc
threadpoolctl
tifffile
torch
torchvision
tqdm
typer
typer-slim
typing-inspection
typing_extensions
tzdata
tzlocal
urllib3
wasabi
weasel
wrapt
